https://github.com/tejaswinibandaru/Samples

Conflicts:
	JpaProject/.classpath
	JpaProject/src/main/java/com/cg/jpaproject/dao/UserDaoImpl.java
	JpaProject/src/main/java/com/cg/jpaproject/service/UserServiceImpl.java
	JpaProject/src/main/java/com/cg/jpaproject/ui/TestApplication.java
	JpaProject/target/classes/com/cg/jpaproject/dao/UserDao.class
	JpaProject/target/classes/com/cg/jpaproject/dao/UserDaoImpl.class
	JpaProject/target/classes/com/cg/jpaproject/service/UserService.class
	JpaProject/target/classes/com/cg/jpaproject/service/UserServiceImpl.class
	JpaProject/target/classes/com/cg/jpaproject/ui/BusReservationSystem.class
	JpaProject/target/classes/com/cg/jpaproject/ui/TestApplication.class

+++ JpaProject/src/main/java/com/cg/jpaproject/service/Validation.java
@@ -195,10 +195,10 @@ public class Validation {
 		return choice;
 	}
 	
-	public BigInteger validateBigIntegerChoice(String input) throws BusException {
+	public Integer validateBigIntegerChoice(String input) throws BusException {
 		  if(!isValidNumeric(input)) 
 			throw new BusException("Integer choice to be entered");
-		  return BigInteger.valueOf(Integer.parseInt(input));
+		  return Integer.parseInt(input);
 		}
 
 }
